skills/wechat-workflow/converting-to-wechat/scripts/requirements.txt
mistune>=3.0.0
PyYAML>=6.0
oss2>=2.18.0
